Fold lightweight free-function properties into function_properties.hpp
